==> 02_notebooks/02_encoding_score.ipynb <==
import pandas as pd

# 분할된 데이터 불러오기
amazon_split_train = pd.read_csv("../01_data/02_processed/sandbox/splits/X_train.csv")
amazon_split_test = pd.read_csv("../01_data/02_processed/sandbox/splits/X_test.csv")
amazon_split_train.head()

# 분할된 데이터 각각에 임시 이름 붙이기 (df1 = train, df2 = test)
df1 = amazon_split_train.copy()
df2 = amazon_split_test.copy()

# 교통 상황과 날씨 변수를 점수화
traffic_map = {"Low":1, "Medium":2, "High":3, "Jam":4}
weather_map = {"Sunny":1, "Cloudy":2, "Windy":3, "Fog":4, "Stormy":5, "Sandstorms":6}

df1["Traffic_score"] = df1["Traffic"].map(traffic_map)
df2["Traffic_score"] = df2["Traffic"].map(traffic_map)

df1["Weather_score"] = df1["Weather"].map(weather_map)
df2["Weather_score"] = df2["Weather"].map(weather_map)

# 원본 열 제거
df1 = df1.drop(["Traffic", "Weather"], axis=1)
df2 = df2.drop(["Traffic", "Weather"], axis=1)
df1.head(50000)

# 결측치를 제거하고 인덱스 번호 맞추기
df1 = df1.dropna(subset=["Traffic_score", "Weather_score"]).reset_index(drop=True)
df2 = df2.dropna(subset=["Traffic_score", "Weather_score"]).reset_index(drop=True)
df1.head(50000)

# score_all 데이터 셋 저장
df1.to_csv("../01_data/02_processed/sandbox/variants/X_train_score_all.csv", index=False, encoding="utf-8-sig")
df2.to_csv("../01_data/02_processed/sandbox/variants/X_test_score_all.csv", index=False, encoding="utf-8-sig")

# 필요한 열만 선택
df3 = df1[["Distance", "Pickup_Wait_Time", "Traffic_score", "Weather_score", "Vehicle_scooter", 
           "Vehicle_van", "Area_Metropolitian", "Area_Other", "Area_Semi_Urban"]].copy()
df4 = df2[["Distance", "Pickup_Wait_Time", "Traffic_score", "Weather_score", "Vehicle_scooter", 
           "Vehicle_van", "Area_Metropolitian", "Area_Other", "Area_Semi_Urban"]].copy()
df3.head(50000)

# score_structural 데이터 셋 저장
df3.to_csv("../01_data/02_processed/sandbox/variants/X_train_score_structural.csv", index=False, encoding="utf-8-sig")
df4.to_csv("../01_data/02_processed/sandbox/variants/X_test_score_structural.csv", index=False, encoding="utf-8-sig")
