==> 02_notebooks/03_encoding_onehot.ipynb <==
import pandas as pd

# 분할된 데이터 불러오기
amazon_split_train = pd.read_csv("../01_data/02_processed/sandbox/splits/X_train.csv")
amazon_split_test = pd.read_csv("../01_data/02_processed/sandbox/splits/X_test.csv")
amazon_split_train.head()

# 분할된 데이터 각각에 임시 이름 붙이기 (df1 = train, df2 = test)
df1 = amazon_split_train.copy()
df2 = amazon_split_test.copy()

## train 데이터 원-핫 인코딩
# 날씨 변수 원-핫 인코딩
# 기준 범주(Sunny) 제거
weather_dummies = pd.get_dummies(df1["Weather"], prefix = "Weather")
weather_dummies = weather_dummies.drop(columns = ["Weather_Sunny"])

# 교통 상황 변수 원-핫 인코딩
# 기준 범주(Low) 제거
traffic_dummies = pd.get_dummies(df1["Traffic"], prefix = "Traffic")
traffic_dummies = traffic_dummies.drop(columns = ["Traffic_Low"])

df1 = pd.concat([df1, weather_dummies, traffic_dummies], axis = 1)
df1.head(50000)

## test 데이터 원-핫 인코딩
weather_dummies = pd.get_dummies(df2["Weather"], prefix = "Weather")
weather_dummies = weather_dummies.drop(columns = ["Weather_Sunny"])

traffic_dummies = pd.get_dummies(df2["Traffic"], prefix = "Traffic")
traffic_dummies = traffic_dummies.drop(columns = ["Traffic_Low"])

df2 = pd.concat([df2, weather_dummies, traffic_dummies], axis = 1)
df2.head(50000)

# 인덱스 번호 맞추기
df1 = df1.dropna().reset_index(drop=True)
df2 = df2.dropna().reset_index(drop=True)

# 원본 열 제거
df1 = df1.drop(["Traffic", "Weather"], axis=1)
df2 = df2.drop(["Traffic", "Weather"], axis=1)
df1.head(50000)

# onehot_all 데이터 셋 저장
df1.to_csv("../01_data/02_processed/sandbox/variants/X_train_onehot_all.csv", index=False, encoding="utf-8-sig")
df2.to_csv("../01_data/02_processed/sandbox/variants/X_test_onehot_all.csv", index=False, encoding="utf-8-sig")

# 필요한 열만 선택
df3 = df1[["Distance", "Pickup_Wait_Time", "Traffic_High", "Traffic_Jam", "Traffic_Medium", "Weather_Cloudy", 
           "Weather_Fog", "Weather_Sandstorms", "Weather_Stormy", "Weather_Windy", "Vehicle_scooter", 
           "Vehicle_van", "Area_Metropolitian", "Area_Other", "Area_Semi_Urban"]].copy()
df4 = df2[["Distance", "Pickup_Wait_Time", "Traffic_High", "Traffic_Jam", "Traffic_Medium", "Weather_Cloudy", 
           "Weather_Fog", "Weather_Sandstorms", "Weather_Stormy", "Weather_Windy", "Vehicle_scooter", 
           "Vehicle_van", "Area_Metropolitian", "Area_Other", "Area_Semi_Urban"]].copy()
df3.head(50000)

# onehot_structural 데이터 셋 저장
df3.to_csv("../01_data/02_processed/sandbox/variants/X_train_onehot_structural.csv", index=False, encoding="utf-8-sig")
df4.to_csv("../01_data/02_processed/sandbox/variants/X_test_onehot_structural.csv", index=False, encoding="utf-8-sig")
